Initial data insertion implemented. Fixed dao for getById and getAll

+++ src/main/java/com/hombre/db/bo/impl/AccountBoImpl.java
@@ -40,7 +40,7 @@ public class AccountBoImpl implements AccountBo {
 	@Override
     @Transactional
 	public Account getAccountById(int id) {
-		return accountDao.getByID((long)id);
+		return accountDao.getByID(id);
 	}
 
 }
